Create API for crawlers in python
An API service, built in Flask, that supports crawling job creation, job
querying, job scheduling (TBI), etc. Service is divided into 2 threads:
one for Flask, and  another for processing the pending jobs queue.
Multithreading is not fully implemented yet; this is because we're
considering moving the API service for crawlers to `go`, because
this is a language that handles concurrency (queue-related operations)
and threading (api & queue) much more beautifully.