Remove __del__
`__del__` method is now deprecated. It has been commented out for a period
of time, so this has no material impact on code functionality.

`__del__` is deprecated because we're unsure when this method can
be called. It is only called when GC is engaged. We're unsure when GC
is working, so this could lead to indeterministic behavior.

For future references, clean-up code can be wrapped in a function, called
when the last intended function is finished.