Migrate JobService posting logic to DBService
When a TVSeries cannot be found in the database, a job should
automatically be scheduled. The job-posting logic was previously
handled by the front end. However, a better approach would be to
delegate that responsibility to the DBService.

The front-end should only know about the DBService; in other words, it
should only know whether or not it can, and has, content to serve to
the user. When a search in DBService fails, it is the DBService's
responsbility to schedule a job at the JobService--not the front-end's.